  \name{usdm-package}

\alias{usdm-package}
\alias{usdm}

\docType{package}

\title{
Uncertainty analysis for SDMs
}

\description{
This package provides a number of functions for exploring the impact of different sources of uncertainties (e.g.positional uncertainty) on performance of species distribution models (SDMs).

In addition, there is a function to quantify different local indicators of spatial association (LISA) for raster data.
}


\author{Babak Naimi \email{naimi.b@gmail.com}


\url{https://r-gis.net/}

\url{https://www.biogeoinformatics.org/}
}

\keyword{ package }
\keyword{ spatial }
